+++ src/js/Servers/FavoriteToggle.jsx
