+++ src/js/Servers/filtersSlice.js
@@ -1,4 +1,4 @@
-import storage from "./../../util/storage.js";
+import storage from "../util/storage.js";
 import { createSlice } from "@reduxjs/toolkit";
 
 const getDefaultState = () => ({
