+++ src/js/TextAnimations.jsx
